third_party/WebKit/LayoutTests/fast/css/pseudo-required-optional-unapplied-expected.txt

Check cases that :required/:optional classes are not applied

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getBackgroundColor("fieldset") is normalColor
PASS getBackgroundColor("legend") is normalColor
PASS getBackgroundColor("label") is normalColor
PASS getBackgroundColor("datalist") is normalColor
PASS getBackgroundColor("optgroup") is normalColor
PASS getBackgroundColor("option") is normalColor
PASS getBackgroundColor("keygen") is normalColor
PASS getBackgroundColor("output") is normalColor
PASS getBackgroundColor("meter") is normalColor
PASS getBackgroundColor("progress") is normalColor
PASS successfullyParsed is true

TEST COMPLETE
